add Main script for modules, delete main functions in modules
