+++ Nika_modules/amino_acids_tools.py
@@ -214,75 +214,3 @@ def summary(sequence: str, record_type: int = 1, percent: bool = False):
     }
     return sum_up
 
-
-def run_aminoacid_seq(sequence: str, function: str = 'summary', record_type: int = 1, percent: bool = False):
-    """
-    Performs the following list of operations:
-    count - counts number of amino acid
-    translate - converts one record type to another
-    determine_charge - counts number or percent of amino acid with different charges
-    determine_polarity - counts number or percent of amino acid with different polarity
-    convert_amino_acid_seq_to_dna - takes an amino acid sequence as input and returns the optimal DNA sequence for E.coli
-    count_possible_number_of_disulfide_bonds - counting the number of possible combinations of two different cysteines to form a disulfide bond
-    count_molecular_weight - takes an amino acid sequence as input and returns the molecular weight of the protein
-    summary - returns results of all functions (default)
-
-    Arguments:
-    - sequence:str - sequence for function
-    - function:str - name of the function you need to perform. You can use: 'count', 'translate', 'summary'(default)
-    - record_type:int - record type of your sequence. You can use 1(default) for single letter type or 3 for three letter type
-    - percent: bool - for determine_charge and determine_polarity shows result in percent (True) or in number (False). Default - False
-
-    Return:
-    - count - int
-    - translate - str
-    - determine_charge - dict
-    - determine_polarity - dict
-    - convert_amino_acid_seq_to_dna - str
-    - count_possible_number_of_disulfide_bonds - int
-    - count_molecular_weight - int
-    - summary - dict
-
-    """
-    if record_type == 1:
-        for amino_acid in sequence:
-            if amino_acid not in single_letter_alphabet:
-                raise ValueError(f"{amino_acid} in your sequence is not amino acid")
-    elif record_type == 3:
-        tuple_sequence = [sequence[i:i + 3] for i in range(0, len(sequence), 3)]
-        for amino_acid in tuple_sequence:
-            if amino_acid not in three_letter_alphabet:
-                raise ValueError(f"{amino_acid} in your sequence is not amino acid")
-        sequence = ''.join(tuple_sequence)
-    if function == 'count':
-        return count(sequence, record_type)
-    elif function == 'translate':
-        return translate(sequence, record_type)
-    elif function == 'summary':
-        return summary(sequence, record_type, percent)
-    elif function == 'determine_polarity':
-        if record_type == 3:
-            return determine_polarity(translate(sequence, record_type), percent)
-        else:
-            return determine_polarity(sequence, percent)
-    elif function == 'determine_charge':
-        if record_type == 3:
-            return determine_charge(translate(sequence, record_type), percent)
-        else:
-            return determine_charge(sequence, percent)
-    elif function == 'count_possible_number_of_disulfide_bonds':
-        if record_type == 3:
-            return count_possible_number_of_disulfide_bonds(translate(sequence, record_type))
-        else:
-            return count_possible_number_of_disulfide_bonds(sequence)
-    elif function == 'count_molecular_weight':
-        if record_type == 3:
-            return count_molecular_weight(translate(sequence, record_type))
-        else:
-            return count_molecular_weight(sequence)
-    elif function == 'convert_amino_acid_seq_to_dna':
-        if record_type == 3:
-            return convert_amino_acid_seq_to_dna(translate(sequence, record_type))
-        else:
-            return convert_amino_acid_seq_to_dna(sequence)
-
